Moved all components out of folders
