+++ frontend/src/components/GroupTab.jsx
