+++ frontend/src/components/PollBox.css
