+++ frontend/src/components/PollBox.jsx
